remove response_containers tests
