+++ test_apps/piccolo_admin_test_app/home/tables.py
@@ -2,7 +2,6 @@ from piccolo.apps.user.tables import BaseUser
 from piccolo.columns import Boolean, ForeignKey, Timestamp, Varchar
 from piccolo.columns.readable import Readable
 from piccolo.table import Table
-
 from piccolo_conf import DB
