+++ tests/response/test_base_response.py
@@ -171,11 +171,6 @@ def test_render_method(body: Any, media_type: MediaType, should_raise: bool) -> 
             assert response.status_code == HTTP_200_OK
 
 
-def test_head_response_doesnt_support_content() -> None:
-    with pytest.raises(ImproperlyConfiguredException):
-        ASGIResponse(body=b"hello world", media_type=MediaType.TEXT, is_head_response=True)
-
-
 def test_get_serializer() -> None:
     class Foo:
         pass
@@ -195,3 +190,8 @@ def test_get_serializer() -> None:
     assert (
         get_serializer(FooResponse(None, type_encoders={Foo: lambda f: "foo"}).response_type_encoders)(Foo()) == "foo"
     )
+
+
+def test_head_response_doesnt_support_content() -> None:
+    with pytest.raises(ImproperlyConfiguredException):
+        ASGIResponse(body=b"hello world", media_type=MediaType.TEXT, is_head_response=True)
